updated code based on reviews
