+++ trailmet/algorithms/distill/__init__.py
@@ -4,4 +4,3 @@ from .factor_transfer import FactorTransfer
 from .losses import KDTransferLoss, RkdDistance, RKdAngle
 from .response_kd import KDTransfer
 from .rkd import RKDTransfer
-from .utils import pdist
