+++ trailmet/utils/__init__.py
@@ -1,3 +1,3 @@
-from .functions import AverageMeter, save_checkpoint, accuracy, seed_everything
+from .functions import AverageMeter, save_checkpoint, accuracy, seed_everything, pdist
 from .benchmark import ModelBenchmark
 from .tp_pruning import TP_Prune
